long

+++ src/loaders/wds_loader.py
@@ -22,7 +22,7 @@ class Collator:
     def _load_data(self, data):
         stream = io.BytesIO(data)
         arr = np.lib.format.read_array(stream)
-        return torch.from_numpy(arr.astype(np.int64))
+        return torch.from_numpy(arr.astype(np.int64)).long()
 
 
     def __call__(self, data):
